refactor: split out classifiers into files
